refactor(signer): avoid redundancies in signature publisher sub modules names
